requirements.txt
Flask==3.0.2
openai==1.14.3
python-dotenv==1.0.1
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.5
